QDoc: Parse Clang-parsed source files explicitly
When QDoc parses a project, it extracts information from its code-base
to obtain and sanitize the user-provided documentation for the project.

When the parsed source file is a C++ file, QDoc uses Clang's APIs to
perform the parsing.

Currently, profiling shows that calling into Clang's APIs by far the
greatest bottleneck, performance-wise, for QDoc.
To reduce the bottleneck, we intend to parallelize calls to the APIs.

Internally, QDoc manages the process of parsing a source-file to the
`CodeParser` family of classes.

For example, `ClangCodeParser`, a child class of `CodeParser`, takes
care of interfacing with Clang's APIs to parse a C++ source file.

The `CodeParser` interface takes care of dispatching to the correct
sub-class for a specific source file and abstracts away the different
access patterns, so that the current structure treats each source-file
equally.

Nonetheless, the `CodeParser` interface and its child classes are not
really suitable for the kind of parallelization work that we intend to
do for `ClangCodeParser`.

Hence, we now separate the source files to parse into Clang-using ones
and ones that do not use Clang's APIs, so that we can use a different
structure when parsing making use of Clang's APIs.

Source files are parsed in the main loop in "main.cpp", by the usage of
`parseSourceFiles`, which uses the abstraction provided by the
`CodeParser` interface to parse the various source files without
source-file specific code.

`parseSourceFile` accepts a `QSet` of source files that are passed on by
the main loop.

As we want to partition the source files into Clang-based and
non-Clang-based ones, the `QSet` argument of `parseSourceFiles` is now
modified to be an `std::vector`, so that we can perform the necessary
order-changing partitioning through `std::(stable_)partition`.

Since the passed in collection of files will now be modified internally
by `parseSourceFiles` and they are not required after they are used
here, the argument was modified to be an rvalue reference to signal the
mutation that the argument receives are scoped to the body of the
function and do not need additional context.

Due to the argument being changed to a vector, some additional code was
added to `parseSourceFiles` to sort and deduplicate the collection of
files, which was previously achieved through the usage of a set.

The calling context of `parseSourceFiles` in the main loop was slightly
modified to conform to the new argument.

`parseSourceFiles` now partitions the collection of source files into
Clang-using source files and non-Clang-using ones.

The loop handling the parsing of the collection of files was split into
two mostly identical ones: The first handling the Clang-based source
files and the second handling all the others.

While the bodies of the split loop are basically identical at the
current point in time, unnecessarily increasing duplication, the body
handling the Clang-based source files will later be changed to allow for
the desired parallel structure.

Task-number: QTBUG-111686
Change-Id: Ic9bd8d2174c6d9ca1ebdbdb94cfc2539a508f3dd
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>