QDoc: Move `outputfromqdocfiles` test project
This change moves the `outputfromqdocfiles` test project from QDoc's
`tst_generatedOutput` to `tst_validateQdocOutputFiles`. The latter is
a data-driven test and better suited for the type of test performed
with this documentation project.

This move is more invasive than those that precede it, in that the test
project is subject to several test cases in `tst_generatedOutput`, that
serve multiple purposes. Of particular interest is how the test project
exercises QDoc's `if(defined)` directive. To achieve test coverage
parity, while at the same time simplify the test project structure, part
of the test is duplicated, and the configuration files for both
resulting cases are consolidated. In addition;

- Merge output format specific configuration from multiple .qdocconf
  files and set their `outputsubdir` such that it aligns with the
  pattern in other test projects.
- Also move the expected output files from `tst_generatedOutput`.
- Remove the cases for all involved configurations in
  `tst_generatedoutput.cpp`.

Task-number: QTBUG-119500
Change-Id: I7868f1935b90f10312671db95c03fffe24c13e75
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>