qdoc: Sort function overloads before generating the documentation
Previously, after reading all the sources, QDoc rearranged the
linked list of function overloads such that the head of the list
pointed to a 'primary' function, i.e. any function that was marked
neither \internal nor \overload. The rest of the overload list
remained in the order source files/function nodes were parsed.

While this provides a stable ordering for the output when parsing
is fully serialized, it became unstable when attempting to
parallelize the process. To address this, sort the list of
overloads.

Instead of attempting to sort a singly-linked list directly,
store them temporarily in a vector, sort them, and reconstruct
the list. Use the previously introduced compare() function
for this purpose. In addition, we need to take account the
status of the functions; whether they are internal, marked
\overload, and whether they are documented.

Internal and undocumented functions are sorted to be at the end
to avoid gaps in the sequence of overload numbers.

Functions not marked \overload take precedence over ones that
are; This is needed because QDoc allows incomplete documentation
for \overload functions.

Note: as a side-effect, this sorting mechanism introduces new
documentation warnings for functions that are not explicitly
marked as \overload, if they were previously selected as an
overload and are now selected as the primary. For these, the
fix is to either document them in full, or mark them as an
\overload if they are not meant as the primary, fully
documented function.

Note: Sections class further sorts copy or move-copy constructors
separately from other constructors; this sorting should be re-thought
and adjusted in a future commit.

Fixes: QTBUG-121481
Change-Id: I975795b49b7b990e6b2e1fca004c5c6c68276ae4
Reviewed-by: Luca Di Sera <luca.disera@qt.io>