QDoc: Ensure correct name for example file
When generating example documentation with DocBook or WebXML as output
format, if `url.examples` isn't set, QDoc attempts to generate
individual pages for each file that's part of the example project. This
can cause QDoc to generate file names that are a canonicalized version
of the path to the file on the system where the documentation is
generated.

The issue was discovered when extending the `tagfiles` test project
such that it shall also generate DocBook and WebXML output. Replacing
calls to `ResolvedFile::get_path()` with `ResolvedFile::get_query()`
in the respective `generateExampleFilePage` implementations avoids
using the canonicalized path to the file as filename, and thus
resolves the issue.

Task-number: QTBUG-119500
Fixes: QTBUG-121774
Change-Id: I09cc400a017328d77428e592213e6506c8cf4161
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>