QDoc: Make CppCodeParser::hasTooManyTopics a free function
QDoc uses a comment-block format to allow the user to provide
documentation specifications that has a concept of "topic commands",
synctactical elements that indicate what the user is intending to
document.

For example, a "\class" command allows the user to specify that they are
writing a certain comment-block with the intention of documenting a C++
record and identify which record they are trying to document.

QDoc has certain rules about what topics commands, and especially how
many, can appear in a single documentation block.

Internally, it uses a method, `CppCodeParser::hasTooManyTopics`, to
identify those comment-blocks that break the required rules, so as to
discard and warn about them.

`hasTooManyTopics` is a method of `CppCodeParser`, the class that
generally takes care of the semantic of certain parts of a QDoc
comment-block.

Consumers of `hasTooManyTopics` generally need an instance of
`CppCodeParser` available to use this method, albeit the method really
has no dependency on any `CppCodeParser`-related state.

`CppCodeParser` is currently part of a legacy interface that is entwined
in QDoc architecture but is not suitable for some of the future
developments, such as certain parallelization efforts, that are planned
for QDoc.

Having `hasTooManyTopics`, being an instance member, adds to a certain
amount of coupling between `CppCodeParser` and other parts of the
code-base that is slowly being removed to enable those developments.

Hence, `hasTooManyTopics` is now moved from being an instance member of
`CppCodeParser` to a free-function.

The declaration of the method in "cppcodeparser.h". The definition of
the method in "cppcodeparser.cpp" was removed from the implementation
file and moved to the relevant header file while being modified to be a
free function.

`hasTooManyTopics` uses certain elements that come from the
"utilities.h" header file and was the only element requiring them in
"cppcodeparser.cpp".

Thus, the inclusion of the "utilities.h" header file was removed from
"cppcodeparser.cpp" and moved to the relevant header file.

The call sites for `hasTooManyTopics` were modified to refer to the
function directly without passing by an instance of `CppCodeParser`.

Change-Id: Ia2d96fe41da24e3d597e311a8abf975f6213aa1b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>