QDoc: Ensure help project QHP SHA1s work cross platform
QDoc can generate Qt Help Project files (`.qhp`-files) through its
`HelpProjectWriter` interface. When it does, it also generates a
`.sha1` file for the generated `.qhp` file. The SHA1 hash it generates
on Windows differs from that it generates on other platforms. This
issue was uncovered when restructuring QDoc's tests for the files it
generates.

Upon investigating the problem, the initial hypothesis was that it was
possibly caused by differing line endings on different platforms. This
theory led to the discovery of a forum.qt.io discussion about a similar
problem:

https://forum.qt.io/topic/124177/qcryptographichash-results-mismatches/3

The forum discussion suggests dropping the `QFile::Text` flag when
opening the file for writing. Indeed, the documentation for
`QIODeviceBase::Text` (the implementation of said flag) states that

    "When writing, the end-of-line terminators are translated
    to the local encoding, for example '\r\n' for Win32."

This means that on Windows, the `QFile::Text` flag can cause line
ending conversions when writing files. Dropping that flag when opening
the output file ensures that the data is written without any automatic
line ending conversions, thus the result should be consistent across
different platforms.

This means QDoc must open the hashFile in binary mode by not passing
the `QFile::Text` flag. The reason for this is that, on Windows, the
`QFile::Text` flag can cause line ending conversions when reading or
writing text files.

Validating the fix is challenging, as `HelpProjectWriter` isn't covered
by any unit tests. However; the change that follows this makes use of
the QHP-generating capabilities of QDoc in a test, and serves to
validate this fix.

Fixes: QTBUG-121850
Task-number: QTBUG-119500
Pick-to: 6.7 6.6 6.5
Change-Id: I9c3c2242c7698e1b7bc031dd33cdbd305d96c713
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>