QDoc: Stop generating .qhp.sha1 files
Since 0f409aca96a6b51271cf86edef948a7981195cf4 (qtbase), QDoc has
generated a .qhp.sha1 file. The purpose of this file was to serve
as an identifier that something in the .qhp file changed, for the
purpose of using the .qhp file as a search index.

However, until 473fbb486926ba8629447d2b91483d982075b4fb, this has
not worked cross-platform, negatively impacting the usefulness of
this feature. Furthermore, there are several types of changes that
do not affect the contents of a .qhp file. That implies that the
usefulness of this hash file is questionable. Finally, obtaining
a SHA1 checksum for the .qhp files is trivial on all three of the
major desktop platforms: use the tool `sha1sum1` on Linux or Mac,
or `CertUtil` on Windows.

Upon further inquiry, it has become evident that the existence of
this file is largely unknown and, indeed, not used anywhere in Qt.
Therefore, this change removes the code reponsible for generating
the .qhp.sha1 files, as well as the QCryptographicHash include.

The commit that introduced this method also `#include`d QDebug,
which is not used in the class. Remove that include as well. This
essentially makes this a revert of the patch that introduced the
feature; as QDoc moved from qtbase.git since then, that part of
the history doesn't belong to qttools.git, effectively enforcing
a new commit instead of a simple revert.

[ChangeLog][QDoc] QDoc no longer generates .qhp.sha1 hash files.

Fixes: QTBUG-121919
Change-Id: I0ad34650109ae03a7f1c3858d31c5027657ab996
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>