QDoc: Handle SharedCommentNodes in processMetaCommands
QDoc allows the user to provide documenation as comment-blocks that are
extracted from a project.

When a comment-block is extracted, it is processed in multiple steps.

The comment-block must be "tied" to a `Node`, an internal representation
for documentable elements.

One way for a comment-block to be "tied" to a `Node` is by the use of
"topic commands", QDoc commands that refer to a documentable element
that QDoc can process to "tie" the block.

For example, the "\fn" command is such a command that the user can use
to refer to a C++ callable element.

When a comment-block is "tied" to a `Node`, certain commands are
processed to modify the "tied" `Node` internal representation.

Those are "meta commands" such as "\relates", which might modify the
hierarchy of the internal representation that for documentable elements
that QDoc has for the project.

This processing is generally performed by "CppCodeParser", which takes
care of managing the semantic of a certain elements of a comment-block.

`CppCodeParser::processTopicArgs` is the entry point for "tying" a
comment-block to a `Node`.

When `processTopicArgs` runs, it processes the various "topic commands"
and returns a list of `Node`s that "tie" to a certain block.

Generally, a comment-block "ties" to a single `Node`. Nontheless, a single
comment block can sometimes, following certain rules, "tie" to more than
one `Node`.

Internally, this is represented by a special type of `Node`,
`SharedCommentNode`, that ties a comment-block to a collection of
`Node`s.

When `processTopicArgs` is run, for those comment-blocks that should be
tied to multiple `Node`s, it creates and return one or more
`SharedCommentNode`s, as required.

While the returned `SharedCommentNode`s that are returned collect all
`Node`s to which the comment-block "tied", the rest of the code does not
necessarily take that into account.

For example, `CppCodeParser::processMetaCommands`, which takes care of
applying "meta commands" to the relevant internal representation that
QDoc uses, treats `SharedCommentNode`s as a single node instead of a
collection.

Due to this lack of processing from other parts of the code-base,
`processTopicArgs` performs a call to `processMetaCommands` for each
`Node` that is part of a created `SharedCommentNode`.

While this generally works, it delinearizes the order and structure in
which comment-blocks are processed.

To streamline this process, the `processMetaCommands` call from
`processTopicArgs` was removed.

The relevant overload of `processMetaCommands` was modified to perform
the application of each "meta command" to all the `Node`s collected by a
`SharedCommentNode`.

The streamlined structure allows the processing of "topic commands" and
"meta commands" to be ordered and batched in the current code-base,
reducing the scope in which certain state is meaningful, which can later
be used to simplify or parallelize certain operations.

Task-number: QTBUG-111686
Change-Id: I5fff3d3618e869b2efb575b45d1ef19fefc25626
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>