QDoc: Move parseFnArg under ClangCodeParser
When QDoc parses the code-base of a project it extracts information
about the code-base itself and the user-provided documentation.

The abstraction taking care of this process is managed by the
`CodeParser` interface, whose subclasses are specialized to handle
certain languages.

For example, `ClangCodeParser` is the subclass taking care of parsing
C++ code through the use of Clang's APIs.

`CodeParser` exposes a virtual method, `parseFnArg`, that is used to
process "\fn" commands in a user-provided documentation block.

This method is used specifically by `ClangCodeParser` only due to some
special handling that is performed for the "\fn" command that requires
access to Clang's APIs.

Since it is only needed by `ClangCodeParser`, the method is now removed
from the `CodeParser` interface and moved directly under
`ClangCodeParser`, where it can be more easily suited to
`ClangCodeParser`-only needs and changes in interface.

The two call-sites of `parseFnArg` were slightly modified to ensure that
a concrete `ClangCodeParser` instance was used to allow access to the
more narrowly scoped method.
To allow for this, an inclusion for "clangcodeparser.h" was added to the
relevant file that contains the call-sites for `parseFnArg` to have
access to a definition for `ClangCodeParser`.

The `override` tag was removed from the declaration of `parseFnArg` in
"clangcodeparser.h" as it is no longer overriding.

Task-number: 111686
Change-Id: I11923d3d6f2b72dea295ed94392e340312ce1446
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>