QDoc: Remove out parameters for processQmlProperties
A user provided documentation block can contain zero or more topic
commands.

A topic command is used to identify which documentable element the
documentation block refers to.

When QDoc parses a user provided documentation block, it registers which
topic commands, if any, are present, in a `Doc`, the internal
representation for documentation blocks, later resolving them to various
`Nodes`, the internal representation for documentable elements.

`CppCodeParser::processTopicArgs` performs this process.

When dealing with QML properties, `processTopicArgs` calls
`CppCodeParser::processQmlProperties`.

`processQmlProperties` accepts two out parameters, a list of `Node`s and a
list of `Doc`s.

When `processQmlProperties` is used, it obtains, as part of its processing,
a series of `Node`s that represent the documentable element that the
process topic refers to.

It stores the found `Node`s into the output parameter, together with the
`Doc` that the topic was produced from, the one the found `Node`s are
"tied" to.

The result stored in those output parameters are later returned as part
of the normal return type, that mimics those output parameters, of `processTopicArgs`.

The output parameters are supposedly there to avoid creating a temporary
that is later merged with the `processTopicArgs` top-level pair of lists.

While this is understandable, the performance hit of doing so is
profiled as meaningless, at the current time, and the use of output
parameters makes it more difficult to update the code for certain
upcoming changes that require changes to the interface itself.

Furthermore, the out parameters generally inflate the actual scope of
the processing that the method performs, making it more difficult to
reason about the elements that are processed.

Hence, to simplify reasoning about the `Node`s and `Doc`s lists produced
in `processQmlProperties` and to reduce the scope of the mutable state that
the output parameters represent, the output parameters are now removed
in favor of returning a value from `processQmlProperties`.

The definition and declaration of `processQmlProperties` were modified to
remove the two parameters. Furthermore, the return type was modified to
be a pair of the required lists.

The definition of `processQmlProperties` was modified to instantiate the two
lists, instead of obtaining them as arguments, further returning them at
the end of the computation.

The single consumer of `processQmlProperties` was modified to stop passing the
two out parameters when calling the method, instead assigning its
returned value to an intermediate value that is merged back with the
top-level pair of lists.

Task-number: QTBUG-111686
Change-Id: Ie3689fb414de4fb960149e8f299703ef648c9a59
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>