QDoc: Long live \qtcmaketargetitem!
QDoc provides a way to document CMake instructions by way of the
`\qtcmakepackage` command. This command is restricted such that it can
only be used at the module level (in other words, in the same comment
that contains the \module command). The command injects the module name
in the CMake instructions that are generated in the header table of
each class reference page that belongs to said module. For most of Qt's
APIs, this works fine. However; there are certain APIs that are
"semi-public", in the sense that they're publically documented while
the symbols belong to a private Qt library (e.g. CorePrivate,
GuiPrivate, etc.). In such cases, the automatically generated
documentation is incorrect.

This patch introduces a new command `\qtcmaketargetitem`. This command
lets users override the target part of the content that is generated
for the `find_package()` CMake instructions. It must be used in
conjunction with the pre-existing `\qtcmakepackage` command.

Fixes: QTBUG-121991
Task-number: QTBUG-121855
Task-number: QTBUG-109214
Change-Id: Ia3ea99d08cd63cdc2f82f2220b6cb5b82557fce1
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>