QDoc: Remove open namespaces
When QDoc obtains the user-provided documentation, it tries to "tie" it
to an internal representation of a documentable-element based on the
"topic commands" of the specific documentation.

For example, a "\class" command, a "\property" command and so on.

When performing this process, while encountering a documentation that
"ties" to certain documentable elements that can be container for other
elements, such as a C++ class or namespace, QDoc stores that
documentable element as an "open namespace".

"open namespaces" are later used when QDoc tries to "tie" a
documentation to an element that might be contained by another element.

For example, if QDoc is trying to "tie" a documentation that has a
"\property" "topic", and it has previously tied a "\class" "topic" to
some C++ class, it will consider the namespace that class provides when
looking for the element that the "\property" refers to.

This allows users to elide fully qualifying certain elements.

For example, if the documentation was trying to refer to a "\property"
`foo` under a class `Bar` that was previously added as an "open
namespace", the user could write a "\property" topic as "\property foo",
with QDoc filling in the `Bar` part of the qualified path `Bar::foo`.

"open namespaces" are currently granular at the file-level, so that all
"open namespaces" that are found by "tying" the documentation block in a
file will be valid only for other documentation blocks in that file.

QDoc performs otherwise shallow searches, currently, so that
non-qualified "topics" would not be found without an "open namespace"
available.

While this allows for a bit more lazyness on the part of the
documentation writer and possibly some minor amount of reducing
repetition, it disables certain properties that are desirable for
optmization changes that are expected to be performed for QDoc.

For example, it reduces the atomicity of a single documentation-block,
creating an ordering between different blocks that prevents
parallelization efforts.

For this same reason, it prevents certain blocks from being freely
moved, for example, as they cannot be resolved in a file that does not
document a suitable "open namespace".

Thus, "open namespaces" are now removed from QDoc.

"open namespaces" are generally stored in `QDocDatabase`, a singleton
class that manages and stores most of the shared mutable state in QDoc.

`QDocDatabase::m_openNamespace`, the member storing the available "open
namespaces" was removed.

`QDocDatabase::clearOpenNamespaces` which was used by consumers of the
"open namespaces" to reset the available "open namespaces" was removed.

`QDoc::insertOpenNamespace` which was used by consumers of the "open
namespaces" to make an "open namespace" available was removed.

`QDocDatabase::findNodeInOpenNamespace` which was used by consumers of
"open namespaces" to "tie" a "topic command" to a documentable entity
while taking into consideration the available "open namespaces" was
removed.

All usages of the above methods were removed.

QDoc regression test suites were updated to take the changes into
account.

Task-number: QTBUG-111686
Change-Id: Id0df9a147b581dadf1fb6a4cf50cd1b5ae71b563
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>