QDoc: Reduce ClangCodeParser dependency on CppCodeParser
QDoc parses the source code of a project to extract the user-provided
documentation and to sanitize it.

For C++ files, it uses Clang's APIs to perform the parsing process.

Profiling shows that calls to Clang are the primary bottleneck with
regards to QDoc performances.

Hence, certain changes, such as parallelizing those calls, are expected
in QDoc.

QDoc uses an interface, `CodeParser`, as the base for components that
perform the source-code parsing process.
For example, `ClangCodeParser` is the `CodeParser` that manages calls to
the Clang's APIs.

`CodeParser` is built in a way that is unsuitable to a parallelization
effort, so that we are moving slowly away from it.

As part of the parsing process, a `CodeParser` extracts a series of
`Doc` elements, an internal representation for documentation-blocks.

Each `Doc` element is processed, so that it can be "tied" to a `Node`,
an internal representation for documentable elements that were extracted
from the code-base, through the use of "topic commands", such as
"\class".

Similarly, each `Doc` element is processes a series of "meta commands",
which changes the internal state of the `Node` the `Doc` is "tied" to.

Both "topic" and "meta" commands are processed by `CppCodeParser`, a
component that is responsible for the semantics of a
documentation-block.

As `CodeParser` expects `Doc` elements to be processed as part of the
parsing of the code-base, all `CodeParser` elements are in some form
dependent on an instance of `CppCodeParser`.

In turn, `CppCodeParser` has a (cyclic) dependency on `ClangCodeParser`,
as resolving the "\fn" topic command currently requires certain calls to
Clang's APIs and certain state that is carried by a `ClangCodeParser`
instance.

In the current code-base, this is solved by the `CodeParser` interface
expecting each `CodeParser` to be a single static instance that is
global and can be reused multiple times.

Due to the current changes in favor of parallelization that are being
done to QDoc, both the cyclic stateful dependency between the two and
the single-instance requirement of the `CodeParser` interface are being
removed, in particular to allow multiple `ClangCodeParser` instances to
be created and to allow parsing multiple C++ source files in parallel.

While due to the very coupled and strict architecture of the
`CodeParser` interface and the general part of the code-base that
perform this parsing process this cannot be solved right away, the patch
performs some initial step to prepare the code-base by reducing the
dependency of `ClangCodeParser` and its usages on the `CodeParser`
interface and by reducing the surface of its dependency to
`CppCodeParser`.

Furthermore, certain operations are reordered to prepare for a
parallelized environment.

`CodeParser::parseSourceFile` is the entry point to the parsing of a
source-file.
It depends on `CppCodeParser` as it expects an implementer of the
`CodeParser` interface to perform the extraction of documentation-blocks
and their processing as a single operation.

While the extraction of documentation-blocks and the general parsing of
a source file is expected to be embarassingly parallel, at least with
regards to `ClangCodeParser`, the processing of documentation blocks is
weak to data-races.

Hence, we aim to split the operations into two different units.

To do so, `ClangCodeParser::parse_cpp_file` is added as an alternative
to `ClangCodeParser::parseSourceFile`.
Due to the code-base not being ready to completely remove the
`CodeParser` interface, the new method lives alongside
`parseSourceFile`, which is modified to be an empty override.

`parse_cpp_file` extracts the documentation-blocks from a source file
but does not process them.
Hence, it does not depend on `CppCodeParser`.

To allow for this, two intermediate representation are added,
`TiedDocumentation` and `UntiedDocumentation`.

When comment-blocks are extracted, they are either "tied" or "to be
tied".

A documentation-block that has no "topic command" is tied directly to the
next documentable element in the source-file.
For example, a documentation-block that precedes a function definition
and has no "\fn" command is tied to the subsequent function definition.
This "tying" process must be performed while the source file is parsed
as it requires the file context itself.

A documentation-block that has at least one "topic command" is instead
"tied" at a later phase, by "CppCodeParser".
This "lazy" "tying" process does not require the source-file context and
can be performed at any later phase.

Previously, this difference was opaque. Since parsing and processing
were performed as a single unit, the two intermediate states of a `Doc`
would not escape the scope of `parseSourceFile`, so that it didn't
require any special handling.

To allow for the split of the parsing and processing step,
`parse_cpp_file` returns a collection of `TiedDocumentation` and
`UntiedDocumentation`, of which each instance represents the "tied" and
"to be tied" case, respectively.

`Doc` that needs "to be tied", are processed by
`CppCodeParser::processTopicArgs`.

Previously, this was done directly in `parseSourceFile`. Now, the
required "tying" calls are performed by the main loop that takes care of
parsing files with `ClangCodeParser`, using the return value of
`parse_cpp_file`.

Similarly, the "processing" step of a `Doc`, by
`CppCodeParser::processMetaCommands`, was moved outside of the parsing
step and is now performed by the main loop after all `Doc` are obtained
and "tied".

The relevant calls to `CppCodeParser` to "tie" and "process" a `Doc`
were modified to work with the new representations, so that the required
state can correctly flow from `parse_cpp_file` to the later phases

The implementation of `PureDocParser::parseSourceFile`, the component
responsible for parsing ".qdoc" files, was slightly modified to respect
the new interface in `CppCodeParser`.

While the changes in `CppCodeParser` are generally equivalent to the
previous code, one change was required due to the change in the ordering
of the operations.

For QML elements, QDoc creates ad-hoc `Node`s in its internal
representation based on the documentation-blocks that are provided by
the user.
This is due to those documentation-blocks being able to appear in C++
and ".qdoc" files and due to QDoc limited support for building an
internal representation for a QML code-base.

For example, when QDoc encounters an "\inqmlmodule" "meta command", it
ties the `Node` that the `Doc` "ties" to so that it is considered a
child of the relevant module.
This is used all-over to allow for internal searches and operations on
the type to be performed correctly based on its new position in the
"hierarchy".

Previously, "parseSourceFile" would "tie" a certain `Doc` and then
instantly "process" the `Doc`.

With the new ordering of operations, all `Doc`s are "tied" before they
are processed.

Certain QML "topic commands" depends on their "tying" `Node` being
positioned correctly under a module.
With "meta commands" being processed after all "topic commands" are
resolved, those QML "topic commands" would fail to resolve correctly,
resulting in missing or misplaced elements in the documentation.

To avoid the issue, QDoc will now reposition a `Node` that represents a
QML type under a module as part of "tying" it, instead of waiting until
the "processing" step, so that the internal `Node` representation has
enough information to resolve the other `Doc` elements.

This change may result in small changes to the resolution of certain
ambiguous links elements, as they are currently dependent on the order
of resolution of certain `Doc` elements.

Task-number: QTBUG-111686
Change-Id: I9a1d602a1e4d2fff35ea4e8de254a72942b302aa
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>