QDoc: Flatten list of headers in main.cpp
When QDoc processes a documentation project it parses the code-base to
extract and sanitize the user-provided documentation.

For projects that contain C++ code, QDoc obtains a series of headers
that need to be parsed based on the project configuration.

Currently, QDoc stores this list of headers in a `QMap` of `QString` to
`QString`, where each key-value pair contains the same element for both
the key and the value.
That is, the map provides sorting and deduplication but uses the
identity function to map key to values, so that it is used as a set.

Hence, the `headers` variable in `processQdocconfFile` in "main.cpp",
that contains the map of headers, is now modified to be a set.

Usages of the variables where modified to conform to the new data
structure.

Change-Id: I598fdad7bf771cce4a5ab159ed1cec584d278943
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>