qdoc: Handle future versions in \deprecated [<version>] command
The \deprecated command accepts a version string as a bracketed argument
representing the version in which the API element was deprecated.

Add a check for whether the argument represents a future version.
The version string is checked against the 'version' .qdocconf variable.
If it's deemed to be a future version, do not mark the Node instance
as deprecated; it will be kept as part of the API reference, instead
of moving it to the separate 'Deprecated members' page.

Tag such nodes with 'until <version>' both in the summary list and
detailed documentation (signature), and add an auto-generated
sentence into the documentation, for example;

  This function is scheduled for deprecation in version 6.9.

Use QVersionNumber for the version comparison. If QVersionNumber cannot
interpret both the 'version' configuration variable and the \deprecated
argument, mark the Node instance as deprecated.

[ChangeLog][QDoc] The \deprecated command now correctly handles
deprecations in a future version.

Fixes: QTBUG-121378
Change-Id: I094699257534902e45edf4fed55cf6ff06021f12
Reviewed-by: Paul Wicking <paul.wicking@qt.io>