QDoc: Separate file parsing and doc processing steps in PureDocParser
QDoc parses the source code of a project to extract the user-provided
documentation and to sanitize it.

QDoc uses an interface, `CodeParser`, as the base for components that
perform the source-code parsing process.
For example, `PureDocParser` is the `CodeParser` that handles ".qdoc"
files.

As part of the parsing process, a `CodeParser` extracts a series of
`Doc` elements, an internal representation for documentation-blocks.

Each `Doc` element is processed, so that it can be "tied" to a `Node`,
an internal representation for documentable elements that were extracted
from the code-base, through the use of "topic commands", such as
"\class".

Similarly, each `Doc` element is processes a series of "meta commands",
which changes the internal state of the `Node` the `Doc` is "tied" to.

Both "topic" and "meta" commands are processed by `CppCodeParser`, a
component that is responsible for the semantics of a
documentation-block.

For dealing with C++ source code, QDoc uses Clang's APIs to perform the
parsing process.

For example, the "topic command" "\fn", to document a C++ callable,
makes underlying calls to Clang APIs when it is being "tied".

Profiling shows that calls to Clang are the primary bottleneck with
regards to QDoc performances.

For each "\fn" commands, QDoc builds a translation unit by calls to
Clang's APIs, which in turn makes for thousand of calls for a project
like Qt, resulting in a noticeable overhead.

Hence, certain changes, such as batching the processing of "\fn"
commands, are being worked on in QDoc.

`CodeParser` is built in a way that is unsuitable to those efforts, so
that we are moving slowly away from it.

`CodeParser` expects `Doc` elements to be processed as part of the
parsing of the code-base. In this sense, each `CodeParser` owns each
`Doc` element from its acquisition the end of its processing step.

"\fn" elements can be obtained from multiple
places, such as ".cpp" and ".qdoc" files, which are handled by different
`CodeParser`s.

To enable the desired batching, we want to collect all "\fn"-owning
`Doc` elements from the various sources and handle them as a collection.

To move towards this goal, we separate the parsing of a source file from
the processing of the `Doc` elements that are extracted from it, so that
we can later reorder and restructure certain operations.

To do so, `PureDocParaser::parse_qdoc_file` is added as an alternative
to `ClangCodeParser::parseSourceFile`.
Due to the code-base not being ready to completely remove the
`CodeParser` interface, the new method lives alongside
`parseSourceFile`, which is modified to be an empty override.

`parse_qdoc_file` extracts the documentation-blocks from a source file
but does not process them.

To allow for this we use a recently added, for a similar purpose,
`UntiedDocumentation` to represent untied and unprocessed `Doc`
elements.

The processing step that is performed by calls to `CppCodeParser` is
moved outside of `parse_qdoc_file`, towards the main loop, and handled
using the newly returned value of `parse_qdoc_file`.

The main loop that handles parsing source files was modified to treat
".qdoc" files specially, so as to be able to follow the new two-step
structure and to call to the non-virtual `parse_qdoc_file`.

This change in the structure currently adds much code duplication, as we
are slowly removing the `CodeParser` abstraction, that is expected to be
revised later on.

Task-number: QTBUG-111686
Change-Id: Iaea67eadeb613564f0265cc39e1bfb40037c6d70
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>