QDoc: Add Config::getExcludedPaths
QDoc uses a singleton `Config` structure to represent and manage the
values defined by a ".qdocconf" file, the configuration file that
defines a project that QDoc can process.

The structure act as a very generic key-value store and generally
provides access only to "raw" values as extracted from a configuration
files.

In turn, the parts of the code that depend on those values generally
perform additional processing to "get them into shape" so that they can
be used.

For certain values, the same processing is performed in multiple part of
the code-base, increasing code-duplication and the opportunity for
error.

For example, a configuration file allow a user to provide a series of
paths that should be excluded from searches that QDoc performs on the
filesystem.

With the current setup, consumers of those excluded paths extract the
"raw" values that the user provided as a list of string, which is then
processed into a set to remove duplicates and provide sorting.

To avoid repeating this processing over multiple parts of the code-base,
a new method, `Config::getExcludedPaths`, was added that return an
already processed set of directories and files to exclude.

The implementation of `Config::getExcludedPaths` performes the same
process that the consumer of those configuration elements did: the list
of excluded files and directories are extracted as raw values from the
configuration so that they can be collected into two sets, which are
used as the "processed" values of those variables.

A record, `ExcludedPaths`, to use a the return value for
`getExcludedPaths`, was added to `Config` to contain the pair of sets
that the "processed" exclusion variables take the form of.

To avoid performing the same process multiple times, considering that
the value of those configuration variables is static, an additional
optional member, `m_excludedPaths`, was added.
`m_excludedPaths` is used to cache the computation between calls to
`getExcludedPaths`.

As `Config` is a singleton that is reused multiple times,
`Config::clear`, which is used when "resetting" the instance to be
reused, was modified to clean the cached value, so that the computation
can be reprocessed when a new configuration file is loaded.

Consumers of the excluded variables were modified to use the new
`getExcludedPaths`.

One of the consumers, `CppCodeParser`, would provides its own caching
mechanism for the processed value of the excluded variables.

As `Config` now performs the caching process, this is not required
anymore.

Hence, `CppCodeParser::m_excludeDirs` and
`CppCodeParser::m_excludeFiles`, the caching member for the excluded
variables in `CppCodeParser`, were removed and substituted with usages
of `Config::getExcludedPaths`.

Task-number: QTBUG-111686
Change-Id: I83f9d3a76cdb9eff70c9cb0596d80072ed4bb3f3
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>